+++ src/navigation/effects.js
@@ -10,6 +10,7 @@ import {
   EmojiAvatar,
   ProfileAvatarSize,
 } from '@/components/asset-list/RecyclerAssetList2/profile-header/ProfileAvatarRow';
+import { HARDWARE_WALLET_TX_NAVIGATOR_SHEET_HEIGHT } from './HardwareWalletTxNavigator';
 
 const statusBarHeight = safeAreaInsetValues.top;
 export const sheetVerticalOffset = statusBarHeight;
@@ -461,6 +462,11 @@ export const expandedPresetWithSmallGestureResponseDistance = {
 export const addWalletNavigatorPreset = ({ route }) => ({
   height: route.params?.sheetHeight,
 });
+
+export const hardwareWalletTxNavigatorPreset = {
+  height: HARDWARE_WALLET_TX_NAVIGATOR_SHEET_HEIGHT,
+  backdropOpacity: 1,
+};
 
 export const sheetPreset = ({ route }) => {
   const shouldUseNonTransparentOverlay =