+++ src/screens/rewards/components/RewardsContent.tsx
@@ -1,26 +1,57 @@
 import React from 'react';
 import { RewardsTitle } from '@/screens/rewards/components/RewardsTitle';
-import { RewardsTotalEarnings } from '@/screens/rewards/components/RewardsTotalEarnings';
-import { RewardsResponseType } from '@/screens/rewards/types/RewardsResponseType';
-import { RewardsPendingEarnings } from '@/screens/rewards/components/RewardsPendingEarnings';
+import { RewardsEarnings } from '@/screens/rewards/components/RewardsEarnings';
+import { RewardsAvailable } from '@/screens/rewards/components/RewardsAvailable';
+import { Rewards } from '@/graphql/__generated__/metadata';
+import { RewardsStats } from './RewardsStats';
+import { RewardsLeaderboard } from '@/screens/rewards/components/RewardsLeaderboard';
+import { RewardsDuneLogo } from '@/screens/rewards/components/RewardsDuneLogo';
+import { ButtonPressAnimation } from '@/components/animations';
 
-type Props = { data: RewardsResponseType };
+const LEADERBOARD_ITEMS_TRESHOLD = 50;
+
+type Props = { data: Rewards };
 
 export const RewardsContent: React.FC<Props> = ({ data }) => {
+  const leaderboardData = data.leaderboard.accounts ?? [];
+  const limitedLeaderboardData = leaderboardData.slice(
+    0,
+    LEADERBOARD_ITEMS_TRESHOLD
+  );
   return (
     <>
       <RewardsTitle text={data.meta.title} />
-      <RewardsTotalEarnings
-        totalEarningsUsd={data.earnings.total.usd}
-        multiplier={data.earnings.multiplier.amount}
-        totalEarningsToken={data.earnings.total.token}
-        tokenImageUrl={data.meta.token.asset.icon_url}
+      {data.earnings && (
+        // TODO: Add explainer sheet navigation to on press here
+        <ButtonPressAnimation onPress={() => {}} scaleTo={0.96}>
+          <RewardsEarnings
+            totalEarnings={data.earnings.total}
+            tokenImageUrl={data.meta.token.asset.iconURL ?? ''}
+            tokenSymbol={data.meta.token.asset.symbol}
+            pendingEarningsToken={data.earnings?.pending.token ?? 0}
+            nextAirdropTimestamp={data.meta.distribution.next}
+            color={data.meta.color}
+          />
+        </ButtonPressAnimation>
+      )}
+      <RewardsAvailable
+        totalAvailableRewards={data.meta.distribution.total}
+        remainingRewards={data.meta.distribution.left}
+        nextDistributionTimestamp={data.meta.distribution.next}
+        color={data.meta.color}
+      />
+      <RewardsStats
+        position={data.stats?.position.current ?? 1}
+        positionChange={data.stats?.position.change.h24 ?? 0}
+        actions={data.stats?.actions ?? []}
+        color={data.meta.color}
+      />
+      <RewardsLeaderboard
+        leaderboard={limitedLeaderboardData}
+        programEndTimestamp={data.meta.end}
         tokenSymbol={data.meta.token.asset.symbol}
       />
-      <RewardsPendingEarnings
-        pendingEarningsUsd={data.earnings.pending.usd}
-        nextAirdropTimestamp={data.meta.next_distribution}
-      />
+      <RewardsDuneLogo />
     </>
   );
 };